+++ src/main/java/com/health/yogiodigym/member/service/MemberService.java
@@ -6,6 +6,7 @@ import com.health.yogiodigym.my.dto.UpdateMemberDto;
 import com.health.yogiodigym.my.dto.UpdateOAuthMemberDto;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import org.springframework.web.multipart.MultipartFile;
 
 import java.util.Set;
 
@@ -29,5 +30,5 @@ public interface MemberService {
 
     void updateOAuthMember(UpdateOAuthMemberDto updateOAuthMemberDto);
 
-    void enrollMaster(Set<String> certificates);
+    void enrollMaster(MultipartFile[] certificate);
 }
