+++ src/main/java/com/health/yogiodigym/my/controller/rest/MyController.java
@@ -6,7 +6,16 @@ import com.health.yogiodigym.common.response.HttpResponse;
 import com.health.yogiodigym.lesson.dto.LessonDto.*;
 import com.health.yogiodigym.lesson.service.LessonService;
 import com.health.yogiodigym.member.entity.MemberOAuth2User;
+import com.health.yogiodigym.member.service.MemberService;
+import com.health.yogiodigym.member.service.NCPStorageService;
+import com.health.yogiodigym.member.service.impl.NCPStorageServiceImpl;
+import com.health.yogiodigym.my.dto.UpdateMemberDto;
+import com.health.yogiodigym.my.dto.UpdateOAuthMemberDto;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -14,15 +23,17 @@ import org.springframework.data.domain.Sort;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
 
+import java.util.HashSet;
 import java.util.List;
+import java.util.Map;
+import java.util.Set;
 
 import static com.health.yogiodigym.common.message.SuccessMessage.*;
 
+@Slf4j
 @RestController
 @RequestMapping("/api/my")
 @RequiredArgsConstructor
@@ -30,6 +41,67 @@ public class MyController {
 
     private final BoardService boardService;
     private final LessonService lessonService;
+    private final MemberService memberService;
+    private final NCPStorageService ncpStorageService;
+
+
+    @PostMapping("/pwd")
+    public ResponseEntity<?> pwd(@RequestBody Map<String, String> checkPwd, @AuthenticationPrincipal MemberOAuth2User principal) {
+        memberService.checkPassword(checkPwd.get("pwd"), principal.getPassword());
+
+        return ResponseEntity.ok().body(new HttpResponse(HttpStatus.OK, PASSWORD_MATCH_SUCCESS.getMessage(), null));
+    }
+
+    @PostMapping("/withdrawal")
+    public ResponseEntity<?> withdrawal(@RequestBody(required = false) Map<String, String> checkPwd, @AuthenticationPrincipal MemberOAuth2User principal,
+                                        HttpServletRequest request, HttpServletResponse response) {
+
+        if(checkPwd != null && !checkPwd.isEmpty()) {
+            memberService.checkPassword(checkPwd.get("pwd"), principal.getPassword());
+        }
+        memberService.registwithdrawal(principal.getMember().getId(), request, response);
+
+        return ResponseEntity.ok().body(new HttpResponse(HttpStatus.OK, WITHDRAWAL_SUCCESS.getMessage(), null));
+    }
+
+    @PutMapping("/info")
+    public ResponseEntity<?> info(@Valid @RequestBody UpdateMemberDto updateMemberDto, @AuthenticationPrincipal MemberOAuth2User principal,
+                                  HttpServletRequest request, HttpServletResponse response) {
+        log.info("update member: {}", updateMemberDto);
+
+        memberService.updateMember(updateMemberDto);
+        memberService.updateAuthentication(principal.getMember().getEmail(), request, response);
+        return ResponseEntity.ok().body(new HttpResponse(HttpStatus.OK, MEMBER_UPDATE_SUCCESS.getMessage(), null));
+    }
+
+    @PutMapping("/oauthinfo")
+    public ResponseEntity<?> oauthinfo(@Valid @RequestBody UpdateOAuthMemberDto updateOAuthMemberDto, @AuthenticationPrincipal MemberOAuth2User principal,
+                                       HttpServletRequest request, HttpServletResponse response) {
+        log.info("update member: {}", updateOAuthMemberDto);
+
+        memberService.updateOAuthMember(updateOAuthMemberDto);
+        memberService.updateAuthentication(principal.getMember().getEmail(), request, response);
+        return ResponseEntity.ok().body(new HttpResponse(HttpStatus.OK, MEMBER_UPDATE_SUCCESS.getMessage(), null));
+    }
+
+    @PostMapping("/profile")
+    public ResponseEntity<?> profile(@RequestParam(value = "profile") MultipartFile profile,
+                                     @AuthenticationPrincipal MemberOAuth2User principal,
+                                     HttpServletRequest request, HttpServletResponse response){
+        ncpStorageService.deleteImageByUrl(principal.getMember().getProfile());
+        String saveFileURL = ncpStorageService.uploadImage(profile, NCPStorageServiceImpl.DirectoryPath.PROFILE);
+        memberService.updateProfile(saveFileURL, principal.getMember().getId());
+        memberService.updateAuthentication(principal.getMember().getEmail(), request, response);
+
+        return ResponseEntity.ok().body(new HttpResponse(HttpStatus.OK, PROFILE_UPDATE_SUCCESS.getMessage(), null));
+    }
+
+    @PostMapping("/master")
+    public ResponseEntity<?> master(@RequestParam("certificate") MultipartFile[] certificate) {
+        memberService.enrollMaster(certificate);
+
+        return ResponseEntity.ok().body(new HttpResponse(HttpStatus.OK, ENROLL_MASTER_SUCCESS.getMessage(), null));
+    }
 
     @GetMapping("/board")
     public ResponseEntity<?> searchMyBoard(@RequestParam(required = false) String boardKeyword,
@@ -38,7 +110,6 @@ public class MyController {
                                            @RequestParam(defaultValue = "0") int page,
                                            @RequestParam(defaultValue = "10") int size,
                                            @AuthenticationPrincipal MemberOAuth2User loginUser) {
-        System.out.println(loginUser.getMember().getId());
 
         Pageable pageable = PageRequest.of(page, size, Sort.by(Sort.Direction.DESC, "id"));
         Page<BoardDetailDto> boards = boardService.searchMyBoards(loginUser.getMember().getId(), boardKeyword, searchColumn, categories, pageable);
@@ -54,8 +125,6 @@ public class MyController {
                                            @RequestParam(defaultValue = "0") int page,
                                            @RequestParam(defaultValue = "10") int size,
                                            @AuthenticationPrincipal MemberOAuth2User loginUser) {
-
-        System.out.println(loginUser.getMember().getId());
 
         Pageable pageable = PageRequest.of(page, size, Sort.by(Sort.Direction.DESC,"id"));
         Page<LessonSearchDto> lessons = lessonService.searchMyLessons(loginUser.getMember().getId(), lessonKeyword, searchColumn, days, categories, pageable);
