+++ src/main/resources/db/migration/V6__alter_member_to_master.sql
@@ -1,11 +1,16 @@
 ALTER TABLE member_to_master
-    DROP COLUMN file,
-    ADD COLUMN enroll_date DATE NOT NULL,
-    ADD COLUMN approval_status VARCHAR(20) NOT NULL;
+    DROP COLUMN file;
 
-CREATE TABLE certificate (
-    `id` BIGINT AUTO_INCREMENT PRIMARY KEY,
-    `enroll_id` BIGINT NOT NULL,
+ALTER TABLE member_to_master
+    ADD COLUMN enroll_date DATE DEFAULT (CURRENT_DATE) NOT NULL;
+
+ALTER TABLE member_to_master
+    ADD COLUMN approval_status VARCHAR(20) DEFAULT 'WAIT' NOT NULL;
+
+CREATE TABLE certificate
+(
+    `id`        BIGINT AUTO_INCREMENT PRIMARY KEY,
+    `enroll_id` BIGINT       NOT NULL,
     `file_name` VARCHAR(255) NOT NULL,
-    FOREIGN KEY (`enroll_id`) REFERENCES member_to_master(`id`) ON DELETE CASCADE
+    FOREIGN KEY (`enroll_id`) REFERENCES member_to_master (`id`) ON DELETE CASCADE
 );
